+++ apps/front-end/public/locales/hi/translation.json
@@ -180,6 +180,7 @@
   "CAMP_DETAILS": "शिविर का विवरण",
   "CAMP_FRONT_VIEW": "शिविर का सामने का दृश्य",
   "CAMP_LOCATION": "शिविर का स्थान",
+  "CAMP_LOCATION_ADDRESS": "शिविर स्थल का पता",
   "CAMP_PERMISSION": "शिविर संपत्ति के मालिक से अनुमति",
   "CAMP_PROPERTY_FACILITIES_CLEAN_LEARNING_ROOM": "स्वच्छ अध्ययन कक्ष",
   "CAMP_PROPERTY_FACILITIES_CLEAN_SAFE_SURROUNDING": "स्वच्छ सुरक्षित क्षेत्र",
@@ -229,6 +230,8 @@
   "CBOSIGNTRANSFER": "CBO द्वारा प्रमाणित ट्रांसफर सर्टिफिकेट",
   "CHANGE_MY_PASSWORD": "मेरा पासवर्ड बदलें",
   "CHANGE_REQUIRED": "बदल आवश्यक है",
+  "CHANGES_NEEDED": "बदलावों की जरूरत",
+  "CHANGES_REQUIRED": "कृपया आवश्यक सुधार के लिए प्रेरक को सूचित करें",
   "CHECK_DOCUMENTS": "दस्तावेजों की जाँच करें",
   "CHOOSE_LANGUAGE": "भाषा चुनें",
   "CLASSROOM_VIEW": "कक्षा का दृश्य",
@@ -518,6 +521,7 @@
   "GO_BACK": "वापस जाओ",
   "GO_TO_BACK": "वापस जाएं",
   "GOOD": "अच्छा",
+  "GOT_THE_KIT": "किट मिला।",
   "GRADUATION": "ग्रेजुएशन",
   "GRADUATION_CERTIFICATES": "स्नातक प्रमाणपत्र",
   "GRAMPANCHAYAT": "ग्राम पंचायत",
@@ -542,6 +546,7 @@
   "IDENTIFY_BENEFICIARY": "शिक्षार्थी की पहचान करें",
   "IDENTIFY_THE_AG_LEARNER": "शिक्षार्थी की पहचान करें",
   "IN_PROGRESS": "कार्य प्रगति पर है",
+  "INACTIVE_GOVERNMENT_PRIVATE_SCHOOL": "निष्क्रिय सरकारी/निजी स्कूल",
   "INCREASE_YOUR_CHANCES_OF_GETTING_SELECTED": "अपने चुने जाने की संभावना बढ़ाएँ",
   "INTERVIEW": "इंटरव्यू",
   "INTERVIEW_CONDUCTED": "इंटरव्यू आयोजित किया गया",
@@ -553,6 +558,7 @@
   "IP_NAME": "IP का नाम",
   "IPHONE": "आईफोन",
   "IS_THE_JOB_RELATED_TO_TEACHING": "क्या नौकरी शिक्षण से संबंधित है?",
+  "IS_THE_KIT_USEFUL": "क्या किट उपयोगी है?",
   "ITS_TIME_TO_START_MOBILIZING": "मोबिलाइज़ करना शुरू करने का समय आ गया है!",
   "JAN_AADHAAR_CARD": "जन आधार कार्ड",
   "JOB_EXPERIENCE": "नौकरी का अनुभव जोड़ें",
@@ -560,6 +566,7 @@
   "JOURNEY_IN_PROJECT_PRAGATI": "प्रगति प्रोग्राम में यात्रा",
   "JUST_ONE_STEP": "बस एक कदम और आगे बढ़ना है!",
   "KIT": "किट",
+  "KIT_DETAILS": "किट विवरण",
   "KIT_SUGGESTION": "किट के लिए आपके सुझाव",
   "KNOW_PRERAK_DUTIES": "प्रेरक के कर्तव्य को जानें",
   "LACK_OF_STUDY": "परिणाम लंबित है",
@@ -573,6 +580,7 @@
   "LEARNER_ASPIRATION": "शिक्षार्थी की आकांक्षाएं",
   "LEARNER_CONSENT_FORM": "शिक्षार्थी और माता-पिता की सहमति",
   "LEARNER_COUNT": "शिक्षार्थी की गिनती",
+  "LEARNER_DETAILS_FAMILY_CONSENT_LETTERS": "शिक्षार्थी पारिवारिक सहमति पत्रों का विवरण देता है",
   "LEARNER_DISTRIBUTION": "शिक्षार्थी वितरण",
   "LEARNER_NOT_AVAILABLE": "शिक्षार्थी उपलब्ध नहीं हैं।",
   "LEARNER_PROFILE": "शिक्षार्थी की प्रोफाइल",
@@ -664,6 +672,7 @@
   "OFFLINE": "ऑफलाइन",
   "OFFLINE_AADHAAR_VERIFICATION": "ऑफलाइन आधार वेरिफिकेशन",
   "OFFLINE_KYC_MESSAGE": "अगर आधार उनके मोबाइल नंबर से जुड़ा है तो आप ऑफलाइन KYC वेरिफिकेशन कर सकते हैं",
+  "OKYC_VERIFICATION": "OKYC वेरिफिकेशन",
   "OLD": "पुराना",
   "OLD_PRERAK": "पुराना प्रेरक",
   "ON_HOLD": "होल्ड पर",
@@ -713,6 +722,7 @@
   "PHD": "पीएचडी",
   "PHOTO": "फोटो",
   "PHOTOS": "तस्वीरें",
+  "PLEASE_ADD_ORIGINAL_RELATED_MATERIAL": "कृपया फिट में ओरिगामी से संबंधित सामग्री जोड़ें। रंगीन कागज, कैंची, गोंद आदि।",
   "PLEASE_CLEAN_CAMERA_LENSE_AND_STEADY_CAMERA": "कृपया अपने कैमरे का लेंस साफ़ करें, पर्याप्त रोशनी सुनिश्चित करें और स्थिर कैमरे से फ़ोटो लें।",
   "PLEASE_ENTER_A_4_DIGIT_NUMBER": "कृपया 4 अंकों की संख्या दर्ज करें। आपको यह कोड याद रखने की आवश्यकता नहीं है।",
   "PLEASE_ENTER_VALID_10_DIGIT_NUMBER": "कृपया सही १० आंकोकी संख्या दर्ज करें",
@@ -723,6 +733,7 @@
   "PLEASE_MENTION_YOUR_REASON": "कृपया कारण बताएं",
   "PLEASE_MENTION_YOUR_REASON_FOR_REJECTING_THE_CANDIDATE": "कृपया उम्मीदवार को अस्वीकार करने के अपने कारण का उल्लेख करें*:",
   "PLEASE_MENTION_YOUR_REASON_FOR_REVIEWING_THE_CANDIDATE": "कृपया उम्मीदवार की समीक्षा करने के अपने कारण का उल्लेख करें:",
+  "PLEASE_NOTE_THAT_THE_FOLLOWING_CHANGES_ARE_AVAILBLE": "कृपया आवश्यक परिवर्तन के लिए प्रेरक को सूचित करे ",
   "PLEASE_REACTIVATE_THE_LEARNER_TO_ACCESS_THE_DETAILS_TAB": "विवरण टैब में डेटा दर्ज करने के लिए शिक्षार्थी को पुन: सक्रिय करें",
   "PLEASE_SELECT": "कृपया चयन करें",
   "PLEASE_SELECT_A_LEARNER": "कृपया शिक्षार्थी चुनें",
@@ -776,6 +787,7 @@
   "PROFILE_DETAILS": "प्रोफ़ाइल विवरण",
   "PROFILE_NEEDS_TO_BE_COMPLETED": "प्रोफ़ाइल को पूरा करें",
   "PROJECT_PRAGATI": "प्रोजेक्ट प्रगति",
+  "PROPERTY_AND_FACILITY_DETAILS": "संपत्ति और सुविधाओं का विवरण",
   "PROPERTY_DETAILS": "संपत्ति का विवरण",
   "PROPERTY_TYPE": "शिविर स्थल का प्रकार ",
   "PURSUING_GRADUATION": "ग्रेजुएशन कर रहे हैं",
@@ -927,6 +939,7 @@
   "SUBJECTS": "विषयों",
   "SUBMIT": "प्रस्तुत करें",
   "SUBMIT_FOR_REGISTRATION": "रजिस्ट्रेशन के लिए प्रस्तुत करें",
+  "SUGGESTIONS_FOR_THE_KIT": "किट के लिए सुझाव",
   "SUPPORT_FOR_EXPLAINING_THE_IMPORTANCE_OF_EDUCATION_TO_THE_FAMILY_MEMBERS": "घर वालों को पढाई का महत्व समझाने में",
   "SUPPORT_FROM_PRAGATI": "प्रगति शिविर से मदद",
   "SUPPORT_IN_BOOKS_AND_AIDS": "किताबें व सहायक सामग्री में मदद",
@@ -952,6 +965,8 @@
   "TELL_US_YOUR_NAME_AS_PER_AADHAAR_CARD": "आधार कार्ड के अनुसार अपना नाम लिखें",
   "TEXT_MESSAGE_MOBILE_NUMBER": "हमने आपको आपके मोबाइल नंबर पर उपयोगकर्ता नाम और पासवर्ड के साथ एक संदेश भेजा है",
   "THE_AGE_OF_THE_LEARNER_SHOULD_BE_15_TO_29_YEARS": "नामांकन के दिन शिक्षार्थी की आयु 14 से 29 वर्ष होनी चाहिए । प्रगति कैम्प में शामिल होने के लिए यह मानदंड जरूरी है ।",
+  "THE_FOLLOWING_FACILITIES_ARE_AVAILABLE_AT_THE_CAMP_SITE": "शिविर स्थल पर निम्नलिखित सुविधाएं उपलब्ध हैं।",
+  "THE_QUALITY_OF_THE_KIT": "किट की गुणवत्ता",
   "THIS_CAN_BE_LETTER_OF": "यह आपकी कंपनी/कार्यस्थल से अनुशंसा पत्र हो सकता है",
   "THIS_CAN_BE_REFERENCE_OR_LETTER_OF": "यह संदर्भ या संगठन से सिफारिश का पत्र हो सकता है",
   "THIS_CAN_BE_YOUR_HIGHEST_GRADE": "यह आपका उच्चतम ग्रेड या स्नातक प्रमाणपत्र हो सकता है",
@@ -977,6 +992,7 @@
   "TRY_AADHAR_UPLOAD_KYC": "आधार फोटो KYC प्रयास करें",
   "TRY_OTHER": "कोई दूसरा कोशिश",
   "TYPE": "प्रकार",
+  "TYPE_OF_CAMPSITE ": "शिविर स्थल का प्रकार ",
   "TYPE_OF_DEVICE": "डिवाइस का प्रकार",
   "TYPE_OF_DOCUMENT": "दस्तावेज़ का प्रकार",
   "TYPE_OF_ENROLLMENT": "नामांकन का प्रकार",
@@ -1030,6 +1046,7 @@
   "VERIFY_AADHAR_NUMER_KYC": "आधार संख्या KYC का उपयोग करके वेरिफिकेशन करें",
   "VERIFY_AADHAR_QR_KYC": "आधार क्यू. आर. KYC को वेरिफिकेशन करें",
   "VERIFY_AADHAR_UPLOAD_KYC": "आधार अपलोड KYC वेरिफिकेशन करें",
+  "VERIFY_MESSAGE": "क्या आप शिविर का सत्यापन करना चाहते हैं? यदि हाँ, तो पुष्टि दबाएँ या शिविर सूची पर वापस जाने के लिए रद्द दबाएँ।",
   "VERIFY_OTP": "वेरिफाई OTP",
   "VIEW": "देखें",
   "VIEW_ALL": "सभी को देखें",
@@ -1074,21 +1091,5 @@
   "YOUR_HIGHEST_QUALIFICATION": "आपका सबसे उच्च शिक्षण",
   "YOUR_IP_WILL_SHARE_DETAILS_SOON": "आपका IP जल्द ही विवरण साझा करेगा",
   "YOUR_PROFILE": "आपका प्रोफ़ाइल",
-  "YOUR_WORK_AVAILABILITY_WILL_BE": "आपके काम की क्या उपलब्धता रहेगी ?",
-  "TYPE_OF_CAMPSITE ": "शिविर स्थल का प्रकार ",
-  "INACTIVE_GOVERNMENT_PRIVATE_SCHOOL": "निष्क्रिय सरकारी/निजी स्कूल",
-  "GOT_THE_KIT": "किट मिला।",
-  "IS_THE_KIT_USEFUL": "क्या किट उपयोगी है?",
-  "SUGGESTIONS_FOR_THE_KIT": "किट के लिए सुझाव",
-  "THE_FOLLOWING_FACILITIES_ARE_AVAILABLE_AT_THE_CAMP_SITE": "शिविर स्थल पर निम्नलिखित सुविधाएं उपलब्ध हैं।",
-  "PLEASE_NOTE_THAT_THE_FOLLOWING_CHANGES_ARE_AVAILBLE": "कृपया आवश्यक परिवर्तन के लिए प्रेरक को सूचित करे ",
-  "PROPERTY_AND_FACILITY_DETAILS": "संपत्ति और सुविधाओं का विवरण",
-  "LEARNER_DETAILS_FAMILY_CONSENT_LETTERS": "शिक्षार्थी पारिवारिक सहमति पत्रों का विवरण देता है",
-  "CAMP_LOCATION_ADDRESS": "शिविर स्थल का पता",
-  "KIT_DETAILS": "किट विवरण",
-  "CHANGES_NEEDED": "बदलावों की जरूरत",
-  "THE_QUALITY_OF_THE_KIT": "किट की गुणवत्ता",
-  "PLEASE_ADD_ORIGINAL_RELATED_MATERIAL": "कृपया फिट में ओरिगामी से संबंधित सामग्री जोड़ें। रंगीन कागज, कैंची, गोंद आदि।",
-  "CHANGES_REQUIRED": "कृपया आवश्यक सुधार के लिए प्रेरक को सूचित करें",
-  "VERIFY_MESSAGE": "क्या आप शिविर का सत्यापन करना चाहते हैं? यदि हाँ, तो पुष्टि दबाएँ या शिविर सूची पर वापस जाने के लिए रद्द दबाएँ।"
+  "YOUR_WORK_AVAILABILITY_WILL_BE": "आपके काम की क्या उपलब्धता रहेगी ?"
 }
