+++ apps/front-end/src/pages/admin/facilitator/view/StatusButton.js
@@ -5,8 +5,10 @@ import {
   facilitatorRegistryService,
   AdminTypo,
   enumRegistryService,
+  checkAadhaar,
 } from "@shiksha/common-lib";
 import { useTranslation } from "react-i18next";
+import AadharCompare from "../../../front-end/AadhaarKyc/AadhaarCompare";
 
 const CRadio = ({ items, onChange }) => {
   const { t } = useTranslation();
@@ -42,7 +44,7 @@ export default function StatusButton({ data, setData }) {
   const [statusList, setStatusList] = React.useState([]);
   const [enumOptions, setEnumOptions] = React.useState({});
   const { t } = useTranslation();
-
+  const subjectId = JSON.parse(data?.program_faciltators?.okyc_response);
   const update = async (status) => {
     if (data?.program_faciltator_id && status) {
       await facilitatorRegistryService.update({
@@ -116,7 +118,7 @@ export default function StatusButton({ data, setData }) {
         ]);
         break;
       case "selected_for_onboarding":
-        setDisabledBtn(["rejected", "quit", "rusticate"]);
+        setDisabledBtn(["rejected", "selected_prerak", "quit", "rusticate"]);
         break;
       case "selected_prerak":
         setDisabledBtn(["rejected", "quit", "rusticate"]);
@@ -179,93 +181,135 @@ export default function StatusButton({ data, setData }) {
           {t(name)}
         </AdminTypo.StatusButton>
       ))}
-      <Modal
-        size={"xl"}
-        isOpen={statusList?.map((e) => e?.name).includes(showModal?.name)}
-        onClose={() => setShowModal()}
-      >
-        <Modal.Content rounded="2xl">
-          <Modal.CloseButton />
-          <Modal.Header borderBottomWidth={0}>
-            <HStack alignItems="center" space={2} justifyContent="center">
-              <AdminTypo.H1 color="textGreyColor.500" bold>
-                {t(showModal?.name)}
-              </AdminTypo.H1>
-            </HStack>
-          </Modal.Header>
-          <Modal.Body pb="5" px="5" pt="0">
-            <VStack space="5">
-              {showModal?.reason ? (
-                <VStack
-                  p="5"
-                  space="5"
-                  flex={1}
-                  borderWidth="1"
-                  borderColor="gray.300"
-                >
-                  <AdminTypo.H6 color="textGreyColor.500" bold>
-                    {t("PLEASE_MENTION_YOUR_REASON")}
-                  </AdminTypo.H6>
-                  <CRadio
-                    onChange={(e) => setReason(e)}
-                    items={
-                      enumOptions[
-                        showModal.status === "quit"
-                          ? "FACILITATOR_REASONS_FOR_QUIT"
-                          : showModal.status === "rusticate"
-                          ? "FACILITATOR_REASONS_FOR_RUSTICATE"
-                          : showModal.status === "rejected"
-                          ? "FACILITATOR_REASONS_FOR_REJECTED"
-                          : showModal.status === "on_hold"
-                          ? "FACILITATOR_REASONS_FOR_ON_REJECTED"
-                          : []
-                      ]
-                    }
-                  />
-                  <Input
-                    onChange={(e) => {
-                      setReason(e?.target?.value);
-                    }}
-                    variant={"underlined"}
-                    placeholder={t("MENTION_YOUR_REASON")}
-                  />
-                </VStack>
-              ) : (
-                <H1 textAlign="center" py="5">
-                  {t("ARE_YOU_SURE")}
-                </H1>
-              )}
-              <HStack width="100%" justifyContent="space-between" space={5}>
-                <AdminTypo.Secondarybutton onPress={() => setShowModal()}>
-                  {t("CANCEL")}
-                </AdminTypo.Secondarybutton>
-                <AdminTypo.PrimaryButton
-                  isDisabled={
-                    !(
-                      (showModal?.reason &&
-                        reason &&
-                        reason?.toLowerCase() != "other") ||
-                      !showModal?.reason
-                    )
-                  }
-                  onPress={() => {
-                    if (
-                      (showModal?.reason &&
-                        reason &&
-                        reason?.toLowerCase() != "other") ||
-                      !showModal?.reason
-                    ) {
-                      update(showModal?.status);
-                    }
-                  }}
-                >
+
+      {showModal?.status !== "selected_prerak" && (
+        <Modal
+          size={"xl"}
+          isOpen={statusList?.map((e) => e?.name).includes(showModal?.name)}
+          onClose={() => setShowModal()}
+        >
+          <Modal.Content rounded="2xl">
+            <Modal.CloseButton />
+            <Modal.Header borderBottomWidth={0}>
+              <HStack alignItems="center" space={2} justifyContent="center">
+                <AdminTypo.H1 color="textGreyColor.500" bold>
                   {t(showModal?.name)}
-                </AdminTypo.PrimaryButton>
+                </AdminTypo.H1>
               </HStack>
-            </VStack>
-          </Modal.Body>
-        </Modal.Content>
-      </Modal>
+            </Modal.Header>
+            <Modal.Body pb="5" px="5" pt="0">
+              <VStack space="5">
+                {showModal?.reason ? (
+                  <VStack
+                    p="5"
+                    space="5"
+                    flex={1}
+                    borderWidth="1"
+                    borderColor="gray.300"
+                  >
+                    <AdminTypo.H6 color="textGreyColor.500" bold>
+                      {t("PLEASE_MENTION_YOUR_REASON")}
+                    </AdminTypo.H6>
+                    <CRadio
+                      onChange={(e) => setReason(e)}
+                      items={
+                        enumOptions[
+                          showModal.status === "quit"
+                            ? "FACILITATOR_REASONS_FOR_QUIT"
+                            : showModal.status === "rusticate"
+                            ? "FACILITATOR_REASONS_FOR_RUSTICATE"
+                            : showModal.status === "rejected"
+                            ? "FACILITATOR_REASONS_FOR_REJECTED"
+                            : showModal.status === "on_hold"
+                            ? "FACILITATOR_REASONS_FOR_ON_REJECTED"
+                            : []
+                        ]
+                      }
+                    />
+                    <Input
+                      onChange={(e) => {
+                        setReason(e?.target?.value);
+                      }}
+                      variant={"underlined"}
+                      placeholder={t("MENTION_YOUR_REASON")}
+                    />
+                  </VStack>
+                ) : (
+                  <H1 textAlign="center" py="5">
+                    {t("ARE_YOU_SURE")}
+                  </H1>
+                )}
+                <HStack width="100%" justifyContent="space-between" space={5}>
+                  <AdminTypo.Secondarybutton onPress={() => setShowModal()}>
+                    {t("CANCEL")}
+                  </AdminTypo.Secondarybutton>
+                  <AdminTypo.PrimaryButton
+                    isDisabled={
+                      !(
+                        (showModal?.reason &&
+                          reason &&
+                          reason?.toLowerCase() != "other") ||
+                        !showModal?.reason
+                      )
+                    }
+                    onPress={() => {
+                      if (
+                        (showModal?.reason &&
+                          reason &&
+                          reason?.toLowerCase() != "other") ||
+                        !showModal?.reason
+                      ) {
+                        update(showModal?.status);
+                      }
+                    }}
+                  >
+                    {t(showModal?.name)}
+                  </AdminTypo.PrimaryButton>
+                </HStack>
+              </VStack>
+            </Modal.Body>
+          </Modal.Content>
+        </Modal>
+      )}
+
+      {showModal?.status == "selected_prerak" && (
+        <Modal
+          isOpen={statusList?.map((e) => e?.name).includes(showModal?.name)}
+          onClose={() => setShowModal()}
+          size={"xl"}
+        >
+          <Modal.CloseButton />
+          <Modal.Content rounded="2xl">
+            <Modal.Body>
+              <VStack space={4}>
+                <AadharCompare
+                  {...{
+                    user: data,
+                    aadhaarCompare: checkAadhaar(data, subjectId?.aadhaar_data),
+                  }}
+                />
+
+                <HStack alignSelf="center" space="4">
+                  <AdminTypo.PrimaryButton
+                    onPress={() => {
+                      update(showModal?.status);
+                    }}
+                  >
+                    {t("CONFIRM")}
+                  </AdminTypo.PrimaryButton>
+                  <AdminTypo.PrimaryButton
+                    onPress={() => {
+                      setShowModal();
+                    }}
+                  >
+                    {t("CHANGE_REQUIRED")}
+                  </AdminTypo.PrimaryButton>
+                </HStack>
+              </VStack>
+            </Modal.Body>
+          </Modal.Content>
+        </Modal>
+      )}
     </Box>
   );
 }
