+++ apps/front-end/src/pages/front-end/facilitator/edit/Form.js
@@ -789,16 +789,14 @@ export default function App({ userTokenInfo, footerLinks }) {
               </FrontEndTypo.Primarybutton>
             ) : (
               <Box>
-                {step !== "aadhaar_details" && (
-                  <FrontEndTypo.Primarybutton
-                    isLoading={loading}
-                    p="4"
-                    mt="4"
-                    onPress={() => onClickSubmit(false)}
-                  >
-                    {t("SAVE_AND_NEXT")}
-                  </FrontEndTypo.Primarybutton>
-                )}
+                <FrontEndTypo.Primarybutton
+                  isLoading={loading}
+                  p="4"
+                  mt="4"
+                  onPress={() => onClickSubmit(false)}
+                >
+                  {t("SAVE_AND_NEXT")}
+                </FrontEndTypo.Primarybutton>
 
                 <FrontEndTypo.Secondarybutton
                   isLoading={loading}
