+++ lib/common-lib/src/components/helper.js
@@ -496,7 +496,7 @@ export const validation = ({ data, key, errors, message, type = '' }) => {
   }
 }
 
-export const checkAadhaar = (user, aadhaar) => {
+export const checkAadhaar = (user, aadhaar, compareList = []) => {
   let isVerified = true
   const arr = [
     'photo',
@@ -515,17 +515,19 @@ export const checkAadhaar = (user, aadhaar) => {
     if (item === 'first_name') {
       result = {
         name: 'FIRST_NAME',
-        aadhaar: first_name,
+        aadhaar: last_name ? first_name : '-',
         user: user?.[item] || '-',
         arr: item
       }
-      if (result?.aadhaar?.toLowerCase() !== result?.user?.toLowerCase()) {
-        isVerified = false
-      }
+      if (compareList?.includes('first_name')) {
+        if (result?.aadhaar?.toLowerCase() !== result?.user?.toLowerCase()) {
+          isVerified = false
+        }
 
-      result = {
-        ...result,
-        isVerified
+        result = {
+          ...result,
+          isVerified
+        }
       }
     } else if (item === 'middle_name') {
       result = {
@@ -534,9 +536,11 @@ export const checkAadhaar = (user, aadhaar) => {
         user: user?.[item] || '-',
         arr: item
       }
-      // if (result?.aadhaar?.toLowerCase() !== result?.user?.toLowerCase()) {
-      //   isVerified = false
-      // }
+      if (compareList?.includes('middle_name')) {
+        if (result?.aadhaar?.toLowerCase() !== result?.user?.toLowerCase()) {
+          isVerified = false
+        }
+      }
     } else if (item === 'last_name') {
       result = {
         name: 'LAST_NAME',
@@ -544,42 +548,50 @@ export const checkAadhaar = (user, aadhaar) => {
         user: user?.[item] || '-',
         arr: item
       }
+
+      if (compareList?.includes('last_name')) {
+        if (result?.aadhaar?.toLowerCase() !== result?.user?.toLowerCase()) {
+          isVerified = false
+        }
+      }
     } else if (item === 'dob') {
       result = {
         name: 'DATE_OF_BIRTH',
-        aadhaar: moment(aadhaar?.dateOfBirth, 'D-M-Y').format('Y-M-D'),
+        aadhaar:aadhaar?.dateOfBirth? moment(aadhaar?.dateOfBirth, 'D-M-Y').format('Y-M-D'):"-",
         user: user?.[item] || '-',
         arr: item
       }
+      if (compareList?.includes('dob')) {
+        const isTrue = moment(result?.user).isSame(result?.aadhaar)
+        if (!isTrue) {
+          isVerified = false
+        }
 
-      const isTrue = moment(result?.user).isSame(result?.aadhaar)
-      if (!isTrue) {
-        isVerified = false
-      }
-
-      result = {
-        ...result,
-        isVerified
+        result = {
+          ...result,
+          isVerified
+        }
       }
     } else if (item === 'gender') {
       result = {
         name: 'GENDER',
-        aadhaar: aadhaar?.gender,
+        aadhaar: aadhaar?.gender?aadhaar?.gender:"-",
         user: user?.[item] || '-',
         arr: item
       }
-
-      /* if (result?.aadhaar === 'M' && result?.user !== 'male') {
-         isVerified = false
-       } else if (result?.aadhaar === 'F' && result?.user !== 'female') {
-         isVerified = false
-       } else if (result?.aadhaar === 'O' && result?.user !== 'other') {
-         isVerified = false
-       }
-       result = {
-         ...result,
-         isVerified
-       } */
+      if (compareList?.includes('dob')) {
+        if (
+          (result?.aadhaar === 'M' && result?.user !== 'male') ||
+          (result?.aadhaar === 'F' && result?.user !== 'female') ||
+          (result?.aadhaar === 'O' && result?.user !== 'other')
+        ) {
+          isVerified = false
+        }
+        result = {
+          ...result,
+          isVerified
+        }
+      }
     } else if (item === 'photo') {
       result = {
         name: 'PHOTO',
@@ -590,19 +602,21 @@ export const checkAadhaar = (user, aadhaar) => {
     } else if (item === 'aadhaar_no') {
       result = {
         name: 'AADHAAR_NO',
-        aadhaar: aadhaar?.maskedNumber,
+        aadhaar: aadhaar?.maskedNumber?aadhaar?.maskedNumber:"-",
         user: user?.aadhar_no,
         arr: item
       }
 
-      if (
-        aadhaar?.maskedNumber?.substr(aadhaar?.maskedNumber?.length - 4) !==
-        result?.user?.substr(result?.user?.length - 4)
-      ) {
-        isVerified = false
-        result = {
-          ...result,
-          isVerified: false
+      if (compareList?.includes('aadhaar_no') || compareList?.length === 0) {
+        if (
+          aadhaar?.maskedNumber?.substr(aadhaar?.maskedNumber?.length - 4) !==
+          result?.user?.substr(result?.user?.length - 4)
+        ) {
+          isVerified = false
+          result = {
+            ...result,
+            isVerified: false
+          }
         }
       }
     }
