+++ lib/common-lib/src/services/facilitatorRegistryService.js
@@ -616,3 +616,27 @@ export const searchByBeneficiary = async (params = {}, header = {}) => {
     return handleResponseException(e)
   }
 }
+
+export const prerakAadhaarOkycResponse = async (params = {}, header = {}) => {
+  try {
+    let headers = {
+      ...header,
+      Authorization: 'Bearer ' + localStorage.getItem('token')
+    }
+    const result = await post(
+      `${process.env.REACT_APP_API_URL}/facilitators/update-okyc-response`,
+      params,
+      {
+        headers: headers
+      }
+    )
+    if (result?.data) {
+      console.log('result?.data', result?.data)
+      return result?.data
+    } else {
+      return {}
+    }
+  } catch (e) {
+    return handleResponseException(e)
+  }
+}
